Divided Honeybee_SplitBuildingMass into two components
